chore: rm Wallet page parallaxes and assets
