+++ src/components/Governance/ParallaxDaoStats/ParallaxDaoStats.tsx
@@ -1,5 +1,5 @@
 import ParallaxDaoStatsElement from '@/components/Governance/ParallaxDaoStats/ParallaxDaoStatsElement'
-import ParallaxText, { type ParallaxTextProps } from '@/components/common/ParallaxTextOld'
+import ParallaxText, { type ParallaxTextProps } from '@/components/common/ParallaxText'
 
 const ParallaxDaoStats = (props: ParallaxTextProps) => {
   return (
