--- src/components/Wallet/MobileAppParallax/styles.module.css
@@ -1,30 +1,0 @@
-.parallaxWrapper {
-  position: relative;
-  text-align: center;
-  width: 200px;
-  margin: 0 auto;
-}
-
-.baseImage {
-  width: 100%;
-  height: auto;
-}
-
-.phone {
-  position: absolute;
-  top: -430px;
-  left: 35px;
-  width: 140px;
-}
-
-@media (min-width: 600px) {
-  .parallaxWrapper {
-    width: 337px;
-  }
-
-  .phone {
-    top: -700px;
-    left: 50px;
-    width: 250px;
-  }
-}
