--- src/components/Wallet/Stepper/styles.module.css
@@ -1,45 +1,0 @@
-.icon {
-  height: 16px;
-  width: 16px;
-  color: var(--mui-palette-primary-light);
-}
-
-.iconActive {
-  color: var(--mui-palette-primary-main);
-}
-
-.stepper :global .MuiStepConnector-line {
-  margin-bottom: -10px;
-  margin-left: -5px;
-  height: 46px;
-  border-left: 2px solid var(--mui-palette-primary-main);
-}
-
-.stepContent {
-  margin-left: 7px;
-  margin-top: -10px;
-  border-left: 2px solid var(--mui-palette-primary-main);
-}
-
-.activeBorder {
-  border-left: 2px dashed var(--mui-palette-primary-light);
-}
-
-.stepper :global .Mui-disabled .MuiStepConnector-line {
-  margin-top: 2px;
-  border-left: 2px dashed var(--mui-palette-primary-light);
-  height: 24px;
-}
-
-/* change color of list item marker */
-.list {
-  list-style: none;
-  padding-left: 0px;
-}
-
-.list li::before {
-  content: '\2022';
-  color: var(--mui-palette-primary-light);
-  display: inline-block;
-  width: 1em;
-}
