+++ src/components/commonCMS/ParallaxText/styles.module.css
@@ -1,15 +1,16 @@
-.textBlock a {
-  text-decoration: underline;
-  color: var(--mui-palette-primary-main);
+.textBlockWrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.textBlockWrapper p {
+  margin: 40px 0;
 }
 
 .textFirst,
 .imageFirst {
   flex-direction: column-reverse;
-}
-
-.textFirstMobile {
-  flex-direction: column;
 }
 
 @media (min-width: 600px) {
